add contactUs on navbar and mobile nav, created style file for each folder NavBar and NavMobileMenu, mobile menu now closes after clicking links, updated CodePurpose title color on NavBar
